requirements.txt
python-dotenv
gspread==6.1.4
google-auth==2.35.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
tabulate==0.9.0
python-gitlab==4.13.0